+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add subcategory filter to grouped bar chart example.
 - Add barHighlight function.
 - Add barFind function to find bars by key.
+- Add barFindByIndex function.
 - Add automatic highlighting of bars on hover to bar series.
 - Add labelHighlight function.
 - Add labelFind function find labels by key.